+++ src/models/index.js
@@ -1,8 +1,10 @@
 const Sequelize = require('sequelize');
 const env = process.env.NODE_ENV || 'development';
-const config = require(__dirname + '/../config/config.json')[env];
+
 const fs = require('fs');
 const path = require('path');
+
+const config = require(path.join(__dirname, '../../config/config.json'))[env];
 
 const db = {};
 const sequelize = new Sequelize(
